@@ -1,18 +1,18 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: ['./index.html', './src/**/*.{vue,js,ts,jsx,tsx}'],
+export default {
+  content: ["./index.html", "./src/**/*.{vue,js,ts,jsx,tsx}"],
   theme: {
     extend: {
       colors: {
-        black: '#191624',
+        black: "#191624",
       },
       animation: {
-        slideup: 'slideup 1s ease-in-out',
-        slidedown: 'slidedown 1s ease-in-out',
-        slideleft: 'slideleft 1s ease-in-out',
-        slideright: 'slideright 1s ease-in-out',
-        wave: 'wave 1.2s linear infinite',
-        slowfade: 'slowfade 2.2s ease-in-out',
+        slideup: "slideup 1s ease-in-out",
+        slidedown: "slidedown 1s ease-in-out",
+        slideleft: "slideleft 1s ease-in-out",
+        slideright: "slideright 1s ease-in-out",
+        wave: "wave 1.2s linear infinite",
+        slowfade: "slowfade 2.2s ease-in-out",
       },
       keyframes: {
         slowfade: {
@@ -20,25 +20,25 @@ module.exports = {
           to: { opacity: 1 },
         },
         slideup: {
-          from: { opacity: 0, transform: 'translateY(25%)' },
-          to: { opacity: 1, transform: 'none' },
+          from: { opacity: 0, transform: "translateY(25%)" },
+          to: { opacity: 1, transform: "none" },
         },
         slidedown: {
-          from: { opacity: 0, transform: 'translateY(-25%)' },
-          to: { opacity: 1, transform: 'none' },
+          from: { opacity: 0, transform: "translateY(-25%)" },
+          to: { opacity: 1, transform: "none" },
         },
         slideleft: {
-          from: { opacity: 0, transform: 'translateX(-20px)' },
-          to: { opacity: 1, transform: 'translateX(0)' },
+          from: { opacity: 0, transform: "translateX(-20px)" },
+          to: { opacity: 1, transform: "translateX(0)" },
         },
         slideright: {
-          from: { opacity: 0, transform: 'translateX(20px)' },
-          to: { opacity: 1, transform: 'translateX(0)' },
+          from: { opacity: 0, transform: "translateX(20px)" },
+          to: { opacity: 1, transform: "translateX(0)" },
         },
         wave: {
-          '0%': { transform: 'scale(0)' },
-          '50%': { transform: 'scale(1)' },
-          '100%': { transform: 'scale(0)' },
+          "0%": { transform: "scale(0)" },
+          "50%": { transform: "scale(1)" },
+          "100%": { transform: "scale(0)" },
         },
       },
     },
